Fixing Some more unit tests and cleaning up code
