+++ waggle-dance-core/pom.xml
@@ -74,23 +74,6 @@
       <scope>provided</scope>
     </dependency>
 
-    <!--&lt;!&ndash; Dropwizard &ndash;&gt;-->
-    <!--<dependency>-->
-      <!--<groupId>io.dropwizard.metrics</groupId>-->
-      <!--<artifactId>metrics-core</artifactId>-->
-      <!--<version>${dropwizard.metrics.version}</version>-->
-    <!--</dependency>-->
-    <!--<dependency>-->
-      <!--<groupId>io.dropwizard.metrics</groupId>-->
-      <!--<artifactId>metrics-graphite</artifactId>-->
-      <!--<version>${dropwizard.metrics.version}</version>-->
-    <!--</dependency>-->
-    <!--<dependency>-->
-      <!--<groupId>io.dropwizard.metrics</groupId>-->
-      <!--<artifactId>metrics-jvm</artifactId>-->
-      <!--<version>${dropwizard.metrics.version}</version>-->
-    <!--</dependency>-->
-
     <!-- Guava -->
     <dependency>
       <groupId>com.google.guava</groupId>
@@ -110,6 +93,12 @@
         <version>1.0.6</version>
     </dependency>
 
+    <dependency>
+        <groupId>io.micrometer</groupId>
+        <artifactId>micrometer-registry-jmx</artifactId>
+        <version>1.0.6</version>
+    </dependency>
+
     <!-- Waggle Dance -->
     <dependency>
       <groupId>com.hotels</groupId>
@@ -121,12 +110,6 @@
       <artifactId>waggle-dance-rest</artifactId>
       <version>${project.version}</version>
     </dependency>
-
-<!--     Hibernate for validation annotations
-    <dependency>
-      <groupId>org.hibernate</groupId>
-      <artifactId>hibernate-validator</artifactId>
-    </dependency> -->
 
     <!-- SSH library -->
     <dependency>
