+++ waggle-dance-integration-tests/src/test/java/com/hotels/bdp/waggledance/metrics/MonitoringConfigurationTestContext.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (C) 2016-2017 Expedia Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hotels.bdp.waggledance.metrics;
+
+import com.google.common.collect.ImmutableMap;
+import com.hotels.bdp.waggledance.conf.GraphiteConfiguration;
+import com.hotels.bdp.waggledance.conf.WaggleDanceConfiguration;
+import io.micrometer.graphite.GraphiteConfig;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MonitoringConfigurationTestContext {
+
+  @Bean
+  GraphiteConfiguration graphiteConfiguration() {
+    GraphiteConfiguration conf = new GraphiteConfiguration();
+    conf.setHost("localhost");
+    conf.setPrefix("graphitePrefix");
+    return conf;
+  }
+
+}
